fix: type updates after refactor
